Group AWS scheduled job constructs
